Added the necessary directories and files to load the generator. The generator.pth is not in the git repo because it's 100MB
In the models\generator\ are the kwargs and the state dict of the pytorch generator. Training directory is where the networks.py is which holds the architecture of the generator and the discriminator. I also added in the utils some of my functions but I also added them in the notebook because it makes more sense to me.